requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
google-auth==1.20.0
google-auth-oauthlib==0.4.1
gspread==3.6.0
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
oauthlib==3.1.0
pathspec==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.1.0
python-dotenv==0.14.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
